added users MongoDB

+++ services/index.js
@@ -1,27 +1,32 @@
 const Contact = require('./schema/contacts');
+const Users = require('./schema/users');
 
-const listContacts = async () => {
-    return Contact.find();
+const listContacts = async (userId) => {
+    return Contact.find({ owner: userId});
 }
 
-const getContactById = async (contactId) => {
-    return Contact.findOne({ _id: contactId });
+const getContactById = async (userId, contactId) => {
+    return Contact.findOne({ owner: userId, _id: contactId });
 };
 
-const addContact = async (contact) => {
-    return Contact.create(contact);
+const addContact = async (userId, contact) => {
+    return Contact.create({ ...contact, owner: userId});
 };
 
-const removeContact = async (contactId) => {
-    return Contact.findOneAndDelete({ _id: contactId });
+const removeContact = async (userId, contactId) => {
+    return await Contact.findOneAndDelete({ owner: userId, _id: contactId }).select({ _id:1 });
 };
 
-const updateContact = async (contactId, body) => {
-    return Contact.findByIdAndUpdate({ _id: contactId }, body, { new:true });
+const updateContact = async (userId, contactId, body) => {
+    return Contact.findByIdAndUpdate({ owner: userId, _id: contactId }, body, { new:true });
 };
 
-const updateStatusContact = async (contactId, body) => {
-    return Contact.findByIdAndUpdate({ _id: contactId }, body, { new: true });
+const updateStatusContact = async (userId, contactId, body) => {
+    return Contact.findByIdAndUpdate({ owner: userId, _id: contactId }, body, { new: true });
+};
+
+const addUser = async (user) => {
+    return Users.create(user);
 };
 
 module.exports = {
@@ -31,4 +36,5 @@ module.exports = {
     removeContact,
     updateContact,
     updateStatusContact,
+    addUser,
 };
